Cleaning up SDL2_net

+++ src/i_net.c
@@ -74,11 +74,11 @@ unsigned short net_to_host16(unsigned int value)
 	return SDLNet_Read16(&s);
 }
 
-unsigned long host_to_net32(unsigned int value)
+unsigned int host_to_net32(unsigned int value)
 {
 	union
 	{
-		unsigned long l;
+		unsigned int l;
 		char b[4];
 	} data;
 
@@ -87,9 +87,9 @@ unsigned long host_to_net32(unsigned int value)
 	return data.l;
 }
 
-unsigned long net_to_host32(unsigned int value)
+unsigned int net_to_host32(unsigned int value)
 {
-	unsigned long l = value;
+	unsigned int l = value;
 
 	return SDLNet_Read32(&l);
 }
